@@ -1,7 +1,7 @@
 package br.serpro.govbr.login
 
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import br.serpro.cidadao.govbr.GovBrWebView
 import br.serpro.cidadao.govbr.GovBrWebViewClient
 import kotlinx.android.synthetic.main.activity_main.*
